Move exception handler to class
